moved Profile.jsx into pages folder
